+++ src/main/java/org/project/soar/model/youthpolicy/repository/YouthPolicyBookmarkRepository.java
@@ -7,6 +7,7 @@ import org.project.soar.model.youthpolicy.YouthPolicyBookmark;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
+import org.springframework.data.domain.Pageable;
 import java.util.List;
 import java.util.Optional;
 
@@ -28,5 +29,29 @@ public interface YouthPolicyBookmarkRepository extends JpaRepository<YouthPolicy
         ORDER BY p.businessPeriodEnd ASC
     """)
     List<YouthPolicy> findLatestBookmarkByEndDate(@Param("user") User user);
+
+    // 전체 인기 (북마크 수 높은 순) — Pageable로 Top N
+    @Query("""
+        SELECT p
+        FROM YouthPolicyBookmark b
+        JOIN b.policy p
+        GROUP BY p
+        ORDER BY COUNT(b) DESC
+    """)
+    List<YouthPolicy> findPopularByBookmarks(Pageable pageable);
+
+    // 나이대별 인기 (북마크 기준) — native (MySQL)
+    @Query(value = """
+        SELECT p.*
+        FROM user_policy_bookmark b
+        JOIN user u ON b.user_id = u.user_id
+        JOIN youth_policy p ON b.policy_id = p.policy_id
+        WHERE u.user_birth_date IS NOT NULL
+          AND FLOOR(TIMESTAMPDIFF(YEAR, u.user_birth_date, CURDATE()) / 10) * 10 = :ageGroup
+        GROUP BY p.policy_id
+        ORDER BY COUNT(*) DESC
+        """, nativeQuery = true)
+    List<YouthPolicy> findPopularBookmarksByAgeGroup(@Param("ageGroup") int ageGroup, Pageable pageable);
+
 
 }
