+++ src/main/java/org/project/soar/model/youthpolicy/service/YouthPolicyBookmarkService.java
@@ -2,7 +2,6 @@ package org.project.soar.model.youthpolicy.service;
 
 import lombok.RequiredArgsConstructor;
 import org.project.soar.model.user.User;
-import org.project.soar.model.user.dto.SignInResponse;
 import org.project.soar.model.user.repository.UserRepository;
 import org.project.soar.model.youthpolicy.YouthPolicy;
 import org.project.soar.model.youthpolicy.YouthPolicyBookmark;
@@ -10,14 +9,15 @@ import org.project.soar.model.youthpolicy.dto.YouthPolicyBookmarkResponseDto;
 import org.project.soar.model.youthpolicy.dto.YouthPolicyLatestResponseDto;
 import org.project.soar.model.youthpolicy.repository.YouthPolicyBookmarkRepository;
 import org.project.soar.model.youthpolicy.repository.YouthPolicyRepository;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
+import org.springframework.data.domain.PageRequest;
 
 @Service
 @RequiredArgsConstructor
@@ -60,6 +60,39 @@ public class YouthPolicyBookmarkService {
         YouthPolicy policy = youthPolicyRepository.findById(policyId)
                 .orElseThrow(() -> new IllegalArgumentException("정책을 찾을 수 없습니다."));
         return bookmarkRepository.findByUserAndPolicy(user, policy).isPresent();
+    }
+
+    /**
+     * 실시간 인기 지원사업
+     */
+    public List<YouthPolicy> getPopularPolicies() {
+        try {
+            return bookmarkRepository.findPopularByBookmarks(PageRequest.of(0, 5));
+        } catch (Exception e) {
+            throw new RuntimeException("실시간 인기 지원사업 조회 중 오류가 발생했습니다.", e);
+        }
+    }
+
+    /**
+     * 나이대별 실시간 인기 지원사업
+     */
+    public List<YouthPolicy> getPopularPoliciesAge(Long userId) {
+        try {
+            User user = userRepository.findByUserId(userId)
+                    .orElseThrow(() -> new RuntimeException("사용자를 찾을 수 없습니다."));
+            // 실제 나이 계산
+            int age = LocalDate.now().getYear() - user.getUserBirthDate().getYear();
+            if (LocalDate.now().getDayOfYear() < user.getUserBirthDate().getDayOfYear()) {
+                age--;
+            }
+            // 10대/20대/30대... 나이대 계산
+            int ageGroup = (age / 10) * 10;
+
+            // 나이대에 맞는 정책 Top 10 조회
+            return bookmarkRepository.findPopularBookmarksByAgeGroup(ageGroup, PageRequest.of(0, 5));
+        } catch (Exception e) {
+            throw new RuntimeException("실시간 인기 지원사업 조회 중 오류가 발생했습니다.", e);
+        }
     }
 
     public YouthPolicyLatestResponseDto getLatestBookmarkByEndDate(Long userId) {
